feat: changes on noTicket file to render tickets options based on api data and files realocated
