+++ src/pages/Dashboard/Payment/index.js
@@ -1,24 +1,26 @@
 import Splash from '../../../components/Splash';
 import useTicket from '../../../hooks/api/useTicket';
 import useEnrollment from '../../../hooks/api/useEnrollment';
-import NoTicket from './NoTicket';
-import NoEnrollment from './NoEnrollment';
+import useTicketTypes from '../../../hooks/api/useTicketType';
 import HasReservedTicket from '../../../components/TicketPayment';
+import NoEnrollment from '../../../components/Ticket/UnathorizedWarning';
+import NoTicket from '../../../components/Ticket';
 
 export default function Payment() {
   const { ticket, ticketLoading, getTicket } = useTicket();
   const { enrollment } = useEnrollment();
+  const { ticketTypes, ticketTypesLoading } = useTicketTypes();
 
   if (!enrollment) {
     return <NoEnrollment />;
   }
 
-  if (ticketLoading) {
+  if (ticketLoading || ticketTypesLoading) {
     return <Splash loading />;
   }
   if (ticket) {
     return <HasReservedTicket ticket={ticket} getTicket={getTicket} />;
   }
 
-  return <NoTicket getTicket={getTicket} />;
+  return <NoTicket getTicket={getTicket} ticketTypes={ticketTypes} />;
 }
