+++ src/services/ticketTypeApi.js
@@ -1,0 +1,11 @@
+import api from './api';
+
+export async function get(token) {
+  const response = await api.get('/tickets/types', {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  });
+
+  return response.data;
+}
